First steps to porting abiword to win64. Still many casts of pointers/handles to 32-bit types, but at least the infrastructure uses now the Win32 API functions that are safe on both win32 and win64 along with defines that are on both systems.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@28467 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/ev/win/ev_Win32Menu.cpp
@@ -370,7 +370,7 @@ bool EV_Win32Menu::synthesizeMenu(XAP_Frame * pFrame, HMENU menuRoot)
 
 					flags |= MF_POPUP;
 					stack.push(sub);
-					u = (UINT) sub;					
+					u = (size_t) sub;					
 					
 					if (szLabelName && *szLabelName)					
 					{
